Add DataUtil fuzzer with encoding corpus for jsoup2

+++ projects/jsoup2/build.sh
@@ -20,10 +20,13 @@ mv $SRC/{*.zip,*.dict} $OUT
 
 # Prepare resources for the DataUtil fuzzer.
 cp -r $SRC/datautil_corpus $OUT/datautil_fuzzer_seed_corpus
-for b64 in $OUT/datautil_fuzzer_seed_corpus/*.b64; do
-  base64 -d "$b64" > "${b64%.b64}"
-  rm "$b64"
-done
+
+if ls "$OUT/datautil_fuzzer_seed_corpus"/*.b64 1> /dev/null 2>&1; then
+  for b64 in "$OUT/datautil_fuzzer_seed_corpus"/*.b64; do
+    base64 -d "$b64" > "${b64%.b64}"
+    rm "$b64"
+  done
+fi
 
 cp $OUT/encodings.dict $OUT/datautil_fuzzer.dict
